+++ api/src/models/db/IBundle.ts
@@ -1,0 +1,10 @@
+export interface IBundle {
+    /**
+     * Id of the bundle
+     */
+    id: string;
+    /**
+     * The transaction hashes contained in the bundle.
+     */
+    transactionHashes: ReadonlyArray<string>;
+}