+++ api/src/models/db/ITransaction.ts
@@ -1,0 +1,10 @@
+export interface ITransaction {
+    /**
+     * Id of the transaction.
+     */
+    id: string;
+    /**
+     * The trytes contained within the transaction.
+     */
+    trytes: string;
+}