+++ api/src/routes/ipfsRetrieve.ts
@@ -1,9 +1,11 @@
-import { composeAPI } from "@iota/core";
-import { asTransactionObject } from "@iota/transaction-converter";
+import { asTransactionObject, Transaction } from "@iota/transaction-converter";
 import { IIPFSRetrieveRequest } from "../models/api/IIPFSRetrieveRequest";
 import { IIPFSRetrieveResponse } from "../models/api/IIPFSRetrieveResponse";
 import { IConfiguration } from "../models/IConfiguration";
 import { IPayload } from "../models/tangle/IPayload";
+import { BundleCacheService } from "../services/bundleCacheService";
+import { TransactionCacheService } from "../services/transactionCacheService";
+import { IotaHelper } from "../utils/iotaHelper";
 import { TrytesHelper } from "../utils/trytesHelper";
 import { ValidationHelper } from "../utils/validationHelper";
 
@@ -17,19 +19,35 @@ export async function ipfsRetrieve(config: IConfiguration, request: IIPFSRetriev
     try {
         ValidationHelper.string(request.transactionHash, "transactionHash");
 
-        const iota = composeAPI({
-            provider: config.node.provider
-        });
+        const hasConnectivity = await IotaHelper.isNodeAvailable(config.node.provider);
 
-        const transactions = await iota.getTrytes([request.transactionHash]);
+        const transactionCacheService = new TransactionCacheService(config.dynamoDbConnection, config.node.provider, hasConnectivity);
+        const transaction = await transactionCacheService.get(request.transactionHash);
 
-        if (!transactions || transactions.length === 0) {
+        if (!transaction) {
             throw new Error(`Unable to locate the specified transaction: '${request.transactionHash}'.`);
         }
 
-        const txObject = asTransactionObject(transactions[0]);
+        const txObject = asTransactionObject(transaction.trytes);
+        const txObjects: Transaction[] = [txObject];
 
-        const payload = TrytesHelper.fromTrytes<IPayload>(txObject.signatureMessageFragment);
+        if (txObject.lastIndex > 0) {
+            const bundleCacheService = new BundleCacheService(config.dynamoDbConnection, config.node.provider, hasConnectivity);
+            const bundle = await bundleCacheService.get(txObject.bundle);
+            if (bundle) {
+                for (let i = 0; i < bundle.transactionHashes.length; i++) {
+                    if (bundle.transactionHashes[i] !== request.transactionHash) {
+                        const txAdditional = await transactionCacheService.get(bundle.transactionHashes[i]);
+                        if (txAdditional) {
+                            txObjects.push(asTransactionObject(txAdditional.trytes));
+                        }
+                    }
+                }
+            }
+            txObjects.sort((a, b) => a.currentIndex - b.currentIndex);
+        }
+
+        const payload = TrytesHelper.fromTrytes<IPayload>(txObjects.map(t => t.signatureMessageFragment).join(""));
 
         return {
             success: true,