+++ api/src/services/stateService.ts
@@ -1,0 +1,17 @@
+import { IState } from "../models/db/IState";
+import { IAWSDynamoDbConfiguration } from "../models/IAWSDynamoDbConfiguration";
+import { AmazonDynamoDbService } from "./amazonDynamoDbService";
+
+/**
+ * Service to handle the state.
+ */
+export class StateService extends AmazonDynamoDbService<IState> {
+    /**
+     * The name of the database table.
+     */
+    public static readonly TABLE_NAME: string = "state";
+
+    constructor(config: IAWSDynamoDbConfiguration) {
+        super(config, StateService.TABLE_NAME, "id");
+    }
+}