+++ api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1.0
+
+* Added transaction/bundle caching layer
+* Added address indexing state
+* Increased storage size to 5Mb
+
 ## v1.0.0
 
-Initial Release
+* Initial Release