+++ client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.1.0
+
+* Increased storage size to 5Mb
+
 ## v1.0.0
 
-Initial Release
+* Initial Release